Move FederationSenderHandler out of generic_worker
